Create cloudformation.yaml
This is that YAML code that deploys a VPC with one public subnet and one private subnet, attaches an Internet Gateway to the public subnet, and a NAT Gateway to the private subnet. It also launches an EC2 instance in the public subnet and an RDS MySQL database in the private subnet.